refactor(agentic): decompose scraping graph into cohesive sub-modules
