+++ backend/scripts/scrape_trending_topics.py
@@ -757,6 +757,56 @@ async def extract_trending_sidebar(
     return topics
 
 
+TWEET_METRICS_EXTRACT_JS = """(selector) => {
+    function parseNum(str) {
+        if (!str) return 0;
+        const match = str.match(/(\\d[\\d,\\.]*\\s*[kKmM]?)/);
+        if (!match) return 0;
+        const clean = match[1].toLowerCase().replace(/,/g, '').trim();
+        if (clean.includes('k')) return Math.round(parseFloat(clean.replace('k', '')) * 1000);
+        if (clean.includes('m')) return Math.round(parseFloat(clean.replace('m', '')) * 1000000);
+        const parsed = parseInt(clean, 10);
+        return isNaN(parsed) ? 0 : parsed;
+    }
+
+    const tweetElements = document.querySelectorAll(selector);
+    const results = [];
+    for (const el of tweetElements) {
+        const textEl = el.querySelector('[data-testid="tweetText"]');
+        const text = textEl ? textEl.innerText : el.innerText;
+        if (!text || text.trim().length === 0) continue;
+
+        const authorEl = el.querySelector('[data-testid="User-Name"]');
+        let author = "unknown";
+        if (authorEl) {
+            const lines = authorEl.innerText.split('\\n');
+            const handleLine = lines.find(l => l.startsWith('@'));
+            author = handleLine || lines[0] || "unknown";
+        }
+
+        const replyBtn = el.querySelector('[data-testid="reply"]');
+        const retweetBtn = el.querySelector('[data-testid="retweet"]');
+        const likeBtn = el.querySelector('[data-testid="like"]');
+        const viewLink = el.querySelector('a[href*="/analytics"]') || el.querySelector('[data-testid="app-text-transition-container"]');
+
+        const replies = parseNum(replyBtn ? replyBtn.innerText || replyBtn.getAttribute('aria-label') : '0');
+        const retweets = parseNum(retweetBtn ? retweetBtn.innerText || retweetBtn.getAttribute('aria-label') : '0');
+        const likes = parseNum(likeBtn ? likeBtn.innerText || likeBtn.getAttribute('aria-label') : '0');
+        const views = parseNum(viewLink ? viewLink.innerText || viewLink.getAttribute('aria-label') : '0');
+
+        results.push({
+            author_handle: author,
+            text: text,
+            replies: replies,
+            retweets: retweets,
+            likes: likes,
+            views: views,
+        });
+    }
+    return results;
+}"""
+
+
 def _parse_evaluated_raw_tweets(
     *, raw_list: Any, dummy_topic_id: uuid.UUID, seen_sigs: set[tuple[str, str]]
 ) -> list[TrendingTweet]:
@@ -813,56 +863,7 @@ async def extract_topic_tweets(
 
     await random_delay(min_sec=0.5, max_sec=1.5)
 
-    extract_js = """(selector) => {
-        function parseNum(str) {
-            if (!str) return 0;
-            const match = str.match(/(\\d[\\d,\\.]*\\s*[kKmM]?)/);
-            if (!match) return 0;
-            const clean = match[1].toLowerCase().replace(/,/g, '').trim();
-            if (clean.includes('k')) return Math.round(parseFloat(clean.replace('k', '')) * 1000);
-            if (clean.includes('m')) return Math.round(parseFloat(clean.replace('m', '')) * 1000000);
-            const parsed = parseInt(clean, 10);
-            return isNaN(parsed) ? 0 : parsed;
-        }
-
-        const tweetElements = document.querySelectorAll(selector);
-        const results = [];
-        for (const el of tweetElements) {
-            const textEl = el.querySelector('[data-testid="tweetText"]');
-            const text = textEl ? textEl.innerText : el.innerText;
-            if (!text || text.trim().length === 0) continue;
-
-            const authorEl = el.querySelector('[data-testid="User-Name"]');
-            let author = "unknown";
-            if (authorEl) {
-                const lines = authorEl.innerText.split('\\n');
-                const handleLine = lines.find(l => l.startsWith('@'));
-                author = handleLine || lines[0] || "unknown";
-            }
-
-            const replyBtn = el.querySelector('[data-testid="reply"]');
-            const retweetBtn = el.querySelector('[data-testid="retweet"]');
-            const likeBtn = el.querySelector('[data-testid="like"]');
-            const viewLink = el.querySelector('a[href*="/analytics"]') || el.querySelector('[data-testid="app-text-transition-container"]');
-
-            const replies = parseNum(replyBtn ? replyBtn.innerText || replyBtn.getAttribute('aria-label') : '0');
-            const retweets = parseNum(retweetBtn ? retweetBtn.innerText || retweetBtn.getAttribute('aria-label') : '0');
-            const likes = parseNum(likeBtn ? likeBtn.innerText || likeBtn.getAttribute('aria-label') : '0');
-            const views = parseNum(viewLink ? viewLink.innerText || viewLink.getAttribute('aria-label') : '0');
-
-            results.push({
-                author_handle: author,
-                text: text,
-                replies: replies,
-                retweets: retweets,
-                likes: likes,
-                views: views,
-            });
-        }
-        return results;
-    }"""
-
-    raw_tweets = await page.evaluate(extract_js, tweet_sel)
+    raw_tweets = await page.evaluate(TWEET_METRICS_EXTRACT_JS, tweet_sel)
     dummy_topic_id = uuid.uuid4()
     seen_sigs: set[tuple[str, str]] = set()
 
@@ -872,12 +873,11 @@ async def extract_topic_tweets(
         seen_sigs=seen_sigs,
     )
 
-    # If fewer than 5 tweets, scroll down smoothly to load more
     if len(tweets) < 5 and hasattr(page, "mouse") and hasattr(page.mouse, "wheel"):
         try:
             await page.mouse.wheel(delta_x=0, delta_y=700)
             await random_delay(min_sec=1.0, max_sec=2.0)
-            more_raw = await page.evaluate(extract_js, tweet_sel)
+            more_raw = await page.evaluate(TWEET_METRICS_EXTRACT_JS, tweet_sel)
             tweets.extend(
                 _parse_evaluated_raw_tweets(
                     raw_list=more_raw,
